+++ alectryon/html.py
@@ -68,7 +68,7 @@ class Gensym():
 
 # pylint: disable=line-too-long
 HEADER = (
-    '<div class="alectryon-header">'
+    '<div class="alectryon-banner">'
     'Built with <a href="https://github.com/cpitclaudel/alectryon/">Alectryon</a>, running {}. '
     'Coq sources are in this panel; goals and messages will appear in the other. '
     'Bubbles (<span class="alectryon-bubble"></span>) indicate interactive fragments: hover for details, tap to reveal contents. '