+++ CHANGES.rst
@@ -5,4 +5,5 @@
 Version 1.1
 ===========
 
+- Rename CSS class ``alectryon-header`` to ``alectryon-banner``.
 - Remove the undocumented ``alectryon-header`` directive.